bnxt_re/lib: Resize CQ support
Adds resize CQ support.
After issuing the resize_cq cmd, library reaps all the
existing completions in the CQ and store it in a list.
The reaping continues till the final CQE is seen. Once
final CQE is seen, lib issues an ibv_cmd_poll_cq so that
the poll_cq context in the kernel is invoked and driver
frees up the resources of the previous CQ.

During a poll CQ from application, the pending completions in
the list is completed first before polling completions from the
new CQ.

Signed-off-by: Selvin Xavier <selvin.xavier@broadcom.com>