bnxt_re: Remove deferred arming logic
Current library arm the CQ in the poll_context. This is based
on the wrong assumption that there will be a poll_cq after a
req_notify_cq.
This may not be true always and notifications stop working if
a poll_cq is not scheduled.

Avoid the deferred arming logic and arm the Hardware CQ from
req_notify_cq itself.

Signed-off-by: Selvin Xavier <selvin.xavier@broadcom.com>